+++ src/main/java/com/klarna/rest/api/customer_token/model/TokenCustomerTokenOrder.java
@@ -30,7 +30,7 @@ import java.util.List;
 /**
  * TokenCustomerTokenOrder
  */
-@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2018-11-02T11:40:30.975Z")
+@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2019-01-25T13:23:40.795Z")
 public class TokenCustomerTokenOrder {
   @JsonProperty("attachment")
   private TokenAttachment attachment = null;